refactor: extract profile helpers
