+++ app/routes/settings+/profile.password.tsx
@@ -27,7 +27,7 @@ import { prisma } from '#app/utils/db.server.ts';
 import { useIsPending } from '#app/utils/misc.tsx';
 import { redirectWithToast } from '#app/utils/toast.server.ts';
 import { PasswordSchema } from '#app/utils/user-validation.ts';
-import { type BreadcrumbHandle } from './profile.tsx';
+import { type BreadcrumbHandle } from './profile-breadcrumbs.tsx';
 
 export const handle: BreadcrumbHandle & SEOHandle = {
   breadcrumb: <Icon name="dots-horizontal">Password</Icon>,
