+++ app/routes/settings+/profile.password_.create.tsx
@@ -21,7 +21,7 @@ import { getPasswordHash, requireUserId } from '#app/utils/auth.server.ts';
 import { prisma } from '#app/utils/db.server.ts';
 import { useIsPending } from '#app/utils/misc.tsx';
 import { PasswordAndConfirmPasswordSchema } from '#app/utils/user-validation.ts';
-import { type BreadcrumbHandle } from './profile.tsx';
+import { type BreadcrumbHandle } from './profile-breadcrumbs.tsx';
 
 export const handle: BreadcrumbHandle & SEOHandle = {
   breadcrumb: <Icon name="dots-horizontal">Password</Icon>,
