+++ app/routes/settings+/profile.two-factor.tsx
@@ -2,7 +2,7 @@ import { type SEOHandle } from '@nasa-gcn/remix-seo';
 import { Outlet } from '@remix-run/react';
 import { Icon } from '#app/components/ui/icon.tsx';
 import { type VerificationTypes } from '#app/routes/_auth+/verify.tsx';
-import { type BreadcrumbHandle } from './profile.tsx';
+import { type BreadcrumbHandle } from './profile-breadcrumbs.tsx';
 
 export const handle: BreadcrumbHandle & SEOHandle = {
   breadcrumb: <Icon name="lock-closed">2FA</Icon>,
