+++ app/routes/settings+/profile.two-factor.verify.tsx
@@ -29,7 +29,7 @@ import { prisma } from '#app/utils/db.server.ts';
 import { getDomainUrl, useIsPending } from '#app/utils/misc.tsx';
 import { redirectWithToast } from '#app/utils/toast.server.ts';
 import { getTOTPAuthUri } from '#app/utils/totp.server.ts';
-import { type BreadcrumbHandle } from './profile.tsx';
+import { type BreadcrumbHandle } from './profile-breadcrumbs.tsx';
 import { twoFAVerificationType } from './profile.two-factor.tsx';
 
 export const handle: BreadcrumbHandle & SEOHandle = {
